feat: return additional context from price check lambda
