+++ amplify/backend/function/pricecheck/src/cardConstants.js
@@ -1,0 +1,255 @@
+const BASIC_LAND_TYPES = [
+  "plains",
+  "island",
+  "islands",
+  "swamp",
+  "swamps",
+  "mountain",
+  "mountains",
+  "forest",
+  "forests",
+  "wastes",
+  "snow-covered plains",
+  "snow-covered island",
+  "snow-covered islands",
+  "snow-covered swamp",
+  "snow-covered swamps",
+  "snow-covered mountain",
+  "snow-covered mountains",
+  "snow-covered forest",
+  "snow-covered forests",
+  "snow-covered wastes",
+];
+
+const LAND_CYCLES = {
+  "cycle-ody-filterland": [
+    "skycloud expanse",
+    "sunscorched divide",
+    "darkwater catacombs",
+    "viridescent bog",
+    "mossfire valley",
+    "ferrous lake",
+    "desolate mire",
+    "shadowblood ridge",
+    "sungrass prairie",
+    "overflowing basin",
+  ],
+  "cycle-checkland": [
+    "glacial fortress",
+    "clifftop retreat",
+    "drowned catacomb",
+    "woodland cemetery",
+    "rootbound crag",
+    "sulfur falls",
+    "isolated chapel",
+    "dragonskull summit",
+    "sunpetal grove",
+    "hinterland harbor",
+  ],
+  "cycle-bfz-tangoland": [
+    "prairie stream",
+    "sunken hollow",
+    "cinder glade",
+    "smoldering marsh",
+    "canopy vista",
+  ],
+  "cycle-horizon-land": [
+    "sunbaked canyon",
+    "nurturing peatland",
+    "fiery islet",
+    "silent clearing",
+    "horizon canopy",
+    "waterlogged grove",
+  ],
+  "cycle-fastland": [
+    "seachrome coast",
+    "inspiring vantage",
+    "darkslick shores",
+    "blooming marsh",
+    "copperline gorge",
+    "spirebluff canal",
+    "concealed courtyard",
+    "blackcleave cliffs",
+    "razorverge thicket",
+    "botanical sanctum",
+  ],
+  "cycle-fetchland": [
+    "flooded strand",
+    "arid mesa",
+    "polluted delta",
+    "verdant catacombs",
+    "wooded foothills",
+    "scalding tarn",
+    "marsh flats",
+    "bloodstained mire",
+    "windswept heath",
+    "misty rainforest",
+  ],
+  "cycle-shm-filterland": [
+    "mystic gate",
+    "rugged prairie",
+    "sunken ruins",
+    "twilight mire",
+    "fire-lit thicket",
+    "cascade bluffs",
+    "fetid heath",
+    "graven cairns",
+    "wooded bastion",
+    "flooded grove",
+  ],
+  "cycle-pathway": [
+    "hengegate pathway",
+    "needleverge pathway",
+    "clearwater pathway",
+    "darkbore pathway",
+    "cragcrown pathway",
+    "riverglide pathway",
+    "brightclimb pathway",
+    "blightstep pathway",
+    "branchloft pathway",
+    "barkchannel pathway",
+  ],
+  "cycle-crowdland": [
+    "sea of clouds",
+    "spectator seating",
+    "morphic pool",
+    "undergrowth stadium",
+    "spire garden",
+    "training center",
+    "vault of champions",
+    "luxury suite",
+    "bountiful promenade",
+    "rejuvenating springs",
+  ],
+  "cycle-painland": [
+    "adarkar wastes",
+    "battlefield forge",
+    "underground river",
+    "llanowar wastes",
+    "karplusan forest",
+    "shivan reef",
+    "caves of koilos",
+    "sulfurous springs",
+    "brushland",
+    "yavimaya coast",
+  ],
+  "cycle-shockland": [
+    "hallowed fountain",
+    "sacred foundry",
+    "watery grave",
+    "overgrown tomb",
+    "stomping ground",
+    "steam vents",
+    "godless shrine",
+    "blood crypt",
+    "temple garden",
+    "breeding pool",
+  ],
+  "cycle-reveal-land": [
+    "port town",
+    "furycalm snarl",
+    "choked estuary",
+    "necroblossom snarl",
+    "game trail",
+    "frostboil snarl",
+    "shineshadow snarl",
+    "foreboding ruins",
+    "fortified village",
+    "vineglimmer snarl",
+  ],
+  "cycle-dual-surveil-land": [
+    "meticulous archive",
+    "elegant parlor",
+    "undercity sewers",
+    "underground mortuary",
+    "commercial district",
+    "thundering falls",
+    "shadowy backstreet",
+    "raucous theater",
+    "lush portico",
+    "hedge maze",
+  ],
+  "tricycle-land": [
+    "indatha triome",
+    "jetmir's garden",
+    "ketria triome",
+    "raffine's tower",
+    "raugrin triome",
+    "savai triome",
+    "spara's headquarters",
+    "xander's lounge",
+    "zagoth triome",
+    "ziatora's proving ground",
+  ],
+  "cycle-triple-tapland": [
+    "arcane sanctum",
+    "crumbling necropolis",
+    "frontier bivouac",
+    "jungle shrine",
+    "mystic monastery",
+    "nomad outpost",
+    "opulent palace",
+    "sandsteppe citadel",
+    "savage lands",
+    "seaside citadel",
+  ],
+  "cycle-abu-dual-land": [
+    "badlands",
+    "bayou",
+    "plateau",
+    "savannah",
+    "scrubland",
+    "taiga",
+    "tropical island",
+    "tundra",
+    "underground sea",
+    "volcanic island",
+  ],
+  "cycle-dsk-thirteenland": [
+    "abandoned campground",
+    "bleeding woods",
+    "etched cornfield",
+    "lakeside shack",
+    "murky sewer",
+    "neglected manor",
+    "peculiar lighthouse",
+    "raucous carnival",
+    "razortrap gorge",
+    "strangled cemetery",
+  ],
+};
+
+const ILLEGAL_SETS = [
+  "30a",
+  "cei",
+  "ced",
+  "wc97",
+  "wc98",
+  "wc99",
+  "wc00",
+  "wc01",
+  "wc02",
+  "wc03",
+  "wc04",
+  "hhO12",
+  "hhO13",
+  "ugl",
+  "unh",
+  "ust",
+];
+
+// Build reverse lookup for LAND_CYCLES
+const CYCLE_LAND_LOOKUP = {};
+for (const oracleTag in LAND_CYCLES) {
+  LAND_CYCLES[oracleTag].forEach((cardName) => {
+    CYCLE_LAND_LOOKUP[cardName.toLowerCase()] = oracleTag;
+  });
+}
+
+// Export the constants
+module.exports = {
+  BASIC_LAND_TYPES,
+  LAND_CYCLES,
+  ILLEGAL_SETS,
+  CYCLE_LAND_LOOKUP,
+};
